Create Publish-Subscribe Code
To receive messages sent to a particular channel, you subscribe to it. When you publish a message to a channel, PubNub delivers that message to everyone subscribed to that channel.

define the message and the channel you're going to send the message to in the publishPayload variable within the publishSampleMessage() function declaration.